Fixed Slicer image "twitching"
Tweaked the angle change calculation so that the slicer sprite will not
constantly flip back and forth due to the high turn ability (which
resulted in overshooting the angle). If turn ability is greater than the
angle difference between the player, slicer will only turn the angle
required (it won't overshoot).