Updated collision sounds
Updated collision sounds so that playership projectiles are responsible
for playing their own collision sounds, the playership plays a collision
sound whenever it is hit, and updated death sound of everything except
sludgermines to be the sludgerdeath sound. This way we also will not
hear a sound every time an enemy collides with something, only when you
collide with something or an enemy dies.