Added Active Powerups to HTML
Added simple Active Powerups section to html. Currently only added it for the double points powerup. In the future the powerups will blink when it is almost out of time.
This active powerups section (and really the rest of the sidebar) is only there until the actual sidebar is created to look more like the original game.