Refactor existing files into module
Refactored existing files into modules as opposed to being loaded as scripts. This will make it easier to break things apart in the future since I am likely going to break apart the fringe file so all of the logic isn't in just one class.
Changing to use modules probably broke Internet Explorer since that is ES 5 syntax, but Internet Explorer is officially not supported by Microsoft anymore so....