more progress on the refactor
Pushing more progress on the refactor so I don't lose my progress. In theory everything should be moved to classes now, all that is left to do is fix the fringe class so that it starts up a game with the new classes. After that just need to clean up all of the TODOs I left for myself, fix bugs I created in the refactor, and remove all of the commented out stuff in the fringe class.