Added way to advance one frame
Added a way to advance the game one frame when the game is paused. This could be useful when debugging to figure out issues. To advance the game one frame when paused hit A. Note that the game must be in debug mode in order of this functionality to be enabled.