Initial HUD Sidebar Info
Added initial HUD sidebar sidebar html and css, hooked in fuel and spare part values as well as score value. Currently level is just a placeholder since level isn't implemented yet, and the systems are not actually implemented yet so all of the circles are colored and the health bar is showing at the bottom. Health and systems will be the next thing to fix, as well as fixing life changes and powerups showing