Finished HUD and Systems!
Finished HUD and Player Systems! While the system damaging could probably be tweaked to be more accurate, it is a good starting point. All damage and health values were multiplied by 4 to account for the about 4-5 times increase in the amount of "player health" since there are 5 systems with "100 health" each.