Updated scanner static effect
Updated the scanner static effect so that the effect could be applied to each sprite separately without effecting other nearby sprites. This does mean that when the effect is happening objects are partially transparent, however looking at gameplay footage that actually appears to be how the old game was as well. Note that the radar context was left alone since the objects on the radar are so small the effects of the overlapping static are negligible.