Allowed both bullet powerups to be active at the same time
Made change to allow both bullet powerups to be active at the same time since that was allowed in the original game. Also started adding some known values pulled from game footage and adding them into the config file so that it is clear that those values are the values that bring the game closest to its original state.